Fixed Bug
v1.-5:
=====
Found a bug where the programme would not remove the period at the end of a "<entity/ies> destroyed <aircraft> flown by <player>" log line. This would result in a user profile for "user"+"." being added.

----------------------------
Changed line 1800:
* std::string get_name = destroyed.substr(0, destroyed.length() - 1);
--> std::string get_name = destroyed.substr(0, destroyed.length() - 2);
----------------------------

If you find any more bugs like this, be sure to report it. Please give the log line that is broken.