Fixed Bug
Found out that the v0.1-5 update that fixed the "<entity/ies> destroyed <aircraft> flown by <player>." bug in Linux broke it in Windows.

Added a branch so that if the system is Linux it will remove the period at the end, but on Windows it will remain the same.

Since this programme sometimes behaves differently on Linux and Windows, if you find any bugs like this please report them. I may have it working on one OS, but it may not be working on the other.