Removed a redundant check
v0.1-6:
Removed a redundant check in the config loading code.
Beforehand, Linux and Windows would load the first two strings differently, but now it is the same.

If you experience any errors, please enable "Debug_Mode" in the config. If you find that some values are not printing out as they should, then please report it.